+++ sdk/src/main/java/com/bugsnag/android/Error.java
@@ -370,6 +370,7 @@ public class Error implements JsonStream.Streamable {
         return config.shouldIgnoreClass(getExceptionName());
     }
 
+    @NonNull
     public HandledState getHandledState() {
         return handledState;
     }